+++ app/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.seoulapp.smartpick"
+        applicationId "com.seoulapp.startpick"
         minSdkVersion 23
         targetSdkVersion 28
         versionCode 1
