+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.seoulapp.smartpick">
+    package="com.seoulapp.startpick">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
@@ -15,10 +15,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.SignupActivity"></activity>
-        <activity android:name=".ui.LoginActivity" />
-        <activity android:name=".ui.SignMainActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name="com.seoulapp.startpick.ui.SignupActivity"></activity>
+        <activity android:name="com.seoulapp.startpick.ui.LoginActivity" />
+        <activity android:name="com.seoulapp.startpick.ui.SignMainActivity" />
+        <activity android:name="com.seoulapp.startpick.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <action android:name="android.intent.action.VIEW" />
