+++ app/src/main/java/com/seoulapp/startpick/data/InfoHomefragData.kt
@@ -1,6 +1,4 @@
-package com.seoulapp.smartpick.data
-
-import android.provider.ContactsContract
+package com.seoulapp.startpick.data
 
 data class InfoHomefragData(
         var image : String,
