+++ app/src/main/java/com/seoulapp/startpick/data/InfoMapfragData.kt
@@ -1,4 +1,4 @@
-package com.seoulapp.smartpick.data
+package com.seoulapp.startpick.data
 
 data class InfoMapfragData(
         var title : String,
