+++ app/src/main/java/com/seoulapp/startpick/ui/JobFragment.kt
@@ -1,14 +1,12 @@
-package com.seoulapp.smartpick.ui
+package com.seoulapp.startpick.ui
 
-import android.content.Context
-import android.net.Uri
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 
-import com.seoulapp.smartpick.R
+import com.seoulapp.startpick.R
 
 
 class JobFragment : Fragment() {
