+++ app/src/main/java/com/seoulapp/startpick/ui/LoginActivity.kt
@@ -1,13 +1,12 @@
-package com.seoulapp.smartpick.ui
+package com.seoulapp.startpick.ui
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import com.seoulapp.smartpick.R
 
-class SignupActivity : AppCompatActivity() {
+class LoginActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_signup)
+        setContentView(R.layout.activity_login)
     }
 }
