+++ app/src/main/java/com/seoulapp/startpick/ui/MapFragment.kt
@@ -1,7 +1,5 @@
-package com.seoulapp.smartpick.ui
+package com.seoulapp.startpick.ui
 
-import android.content.Context
-import android.net.Uri
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.support.v7.widget.LinearLayoutManager
@@ -9,12 +7,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 
-import com.seoulapp.smartpick.R
-import com.seoulapp.smartpick.data.InfoHomefragData
-import com.seoulapp.smartpick.data.InfoMapfragData
-import com.seoulapp.smartpick.ui.adapter.InfoHomeAdapter
-import com.seoulapp.smartpick.ui.adapter.InfoMapAdapter
-import kotlinx.android.synthetic.main.fragment_home.*
+import com.seoulapp.startpick.R
+import com.seoulapp.startpick.data.InfoMapfragData
+import com.seoulapp.startpick.ui.adapter.InfoMapAdapter
 import kotlinx.android.synthetic.main.fragment_map.*
