+++ app/src/main/java/com/seoulapp/startpick/ui/SignupActivity.kt
@@ -1,12 +1,13 @@
-package com.seoulapp.smartpick.ui
+package com.seoulapp.startpick.ui
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.seoulapp.startpick.R
 
-class LoginActivity : AppCompatActivity() {
+class SignupActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
+        setContentView(R.layout.activity_signup)
     }
 }
