+++ app/src/main/java/com/seoulapp/startpick/ui/adapter/InfoMapAdapter.kt
@@ -1,7 +1,6 @@
-package com.seoulapp.smartpick.ui.adapter
+package com.seoulapp.startpick.ui.adapter
 
 import android.content.Context
-import android.support.v7.view.menu.MenuView
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
@@ -9,9 +8,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import com.bumptech.glide.Glide
-import com.seoulapp.smartpick.R
-import com.seoulapp.smartpick.data.InfoHomefragData
-import com.seoulapp.smartpick.data.InfoMapfragData
+import com.seoulapp.startpick.R
+import com.seoulapp.startpick.data.InfoMapfragData
 
 class InfoMapAdapter(private val ctx: Context, private val dataList: ArrayList<InfoMapfragData>) : RecyclerView.Adapter<InfoMapAdapter.Holder>() {
     override fun onCreateViewHolder(viewGroup: ViewGroup, viewtype: Int): Holder {
