+++ app/src/main/res/layout/activity_main.xml
@@ -7,13 +7,13 @@
     android:layout_height="match_parent"
     tools:context="MainActivity">
 
-    <com.seoulapp.smartpick.ui.view.NoneSwipeViewPager
+    <com.seoulapp.startpick.ui.view.NoneSwipeViewPager
         android:id="@+id/activity_main_vp_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1">
 
-    </com.seoulapp.smartpick.ui.view.NoneSwipeViewPager>
+    </com.seoulapp.startpick.ui.view.NoneSwipeViewPager>
 
     <android.support.design.widget.TabLayout
         android:id="@+id/activity_main_tl_navi"
